requirements.txt
flask
python-escpos
Pillow
